+++ CHANGELOG.md
@@ -43,6 +43,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - [#305](https://github.com/babylonlabs-io/babylon/pull/305) chore: add more error logs to `VerifyInclusionProofAndGetHeight`
 - [#304](https://github.com/babylonlabs-io/babylon/pull/304) Add highest voted height to finality provider
 - [#314](https://github.com/babylonlabs-io/babylon/pull/314) Require exact unbonding time in delegation
+- [#317](https://github.com/babylonlabs-io/babylon/pull/317) Enforce that EOI
+delegations using correct parameters version
 
 ### State Machine Breaking