+++ CHANGELOG.md
@@ -51,6 +51,11 @@ delegations using correct parameters version
 - [#310](https://github.com/babylonlabs-io/babylon/pull/310) implement adr-37 -
 making params valid for btc light client ranges
 
+### Bug fixes
+
+- [#318](https://github.com/babylonlabs-io/babylon/pull/318) Fix BTC delegation status check
+to relay on UnbondingTime in delegation
+
 ## v0.17.2
 
 ### Improvements