+++ CHANGELOG.md
@@ -37,6 +37,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+## v0.18.0
+
 ### Improvements
 
 - [#309](https://github.com/babylonlabs-io/babylon/pull/309) feat(adr-036): custom withdrawal address